requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alpaca-py==0.42.2
alpha_vantage==3.0.0
annotated-types==0.7.0
asyncio-nats-client==0.11.5
attrs==25.3.0
backtrader==1.9.78.123
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
Deprecated==1.2.18
deprecation==2.1.0
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.1
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.9.0
greenlet==3.2.4
gunicorn==23.0.0
huggingface-hub==0.35.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
llvmlite==0.44.0
MarkupSafe==3.0.3
matplotlib==3.10.6
mpmath==1.3.0
msgpack==1.0.3
multidict==6.6.4
multitasking==0.0.12
networkx==3.5
newsapi-python==0.2.7
nltk==3.9.2
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pandas-ta==0.4.71b0
peewee==3.18.2
pillow==11.3.0
platformdirs==4.4.0
propcache==0.3.2
protobuf==6.32.1
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.1
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentencepiece==0.2.1
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.43
sseclient-py==1.8.0
sympy==1.14.0
textblob==0.19.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.57.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.20.1
yfinance==0.2.66
